# passport/docs/swagger.yaml
basePath: /v1
definitions:
  gin.Error:
    properties:
      err:
        type: error
      meta:
        type: object
      type:
        type: ErrorType
    type: object
  protocol.AccessTokenRequest:
    properties:
      clientId:
        type: string
      code:
        type: string
      grantType:
        type: string
      redirectURI:
        type: string
    type: object
  protocol.AccessTokenResponse:
    properties:
      accessToken:
        type: string
      expiresIn:
        type: integer
      idToken:
        type: string
      refreshToken:
        type: string
      tokenType:
        type: string
    type: object
  protocol.AuthorizeRequest:
    properties:
      clientId:
        type: string
      redirectURI:
        type: string
      responseType:
        type: string
      scope:
        type: string
      state:
        type: string
    type: object
  protocol.CreateUserCommand:
    properties:
      code:
        description: 验证码
        type: integer
      email:
        description: 邮箱
        type: string
      password:
        description: 密码
        type: string
    type: object
  protocol.PasswordLoginCommand:
    properties:
      email:
        description: 邮箱
        type: string
      password:
        description: 密码
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /oauth2.0/authorize:
    get:
      consumes:
      - application/json
      parameters:
      - description: 授权请求
        in: body
        name: AuthorizeRequest
        required: true
        schema:
          $ref: '#/definitions/protocol.AuthorizeRequest'
      produces:
      - application/json
      responses:
        "302": {}
  /oauth2.0/token:
    post:
      consumes:
      - application/json
      parameters:
      - description: 获取access token
        in: body
        name: AccessTokenRequest
        required: true
        schema:
          $ref: '#/definitions/protocol.AccessTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取token成功
          schema:
            $ref: '#/definitions/protocol.AccessTokenResponse'
  /users:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户注册请求
        in: body
        name: CreateUserCommand
        required: true
        schema:
          $ref: '#/definitions/protocol.CreateUserCommand'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/protocol.AccessTokenResponse'
        "409":
          description: Can not find ID
          schema:
            $ref: '#/definitions/gin.Error'
      summary: 用户注册
  /users:passwordLogin:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户注册请求
        in: body
        name: PasswordLoginCommand
        required: true
        schema:
          $ref: '#/definitions/protocol.PasswordLoginCommand'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功返回认证token
          schema:
            $ref: '#/definitions/protocol.AccessTokenResponse'
        "401":
          description: 密码输入错误
          schema:
            $ref: '#/definitions/gin.Error'
      summary: 账号密码登录
swagger: "2.0"